PoC to increase radio frame length for DECT NR+ Radio
Increase radio frame length to carry the largest data size for DECT NR+ radio on nRF9151.

The largest data size of MCS 4 is 700 bytes per 8 subslots:
https://docs.nordicsemi.com/bundle/ncs-3.0.0/page/nrfxlib/nrf_modem/doc/dectphy.html

subtract 3 bytes header for DECT PHY (1 byte sequence and 2 bytes frame size),
the rest 697 bytes are available for carrying OpenThread MAC frame.

+++ src/core/thread/mesh_forwarder.hpp
@@ -385,7 +385,7 @@ private:
     static constexpr uint8_t kFailedCslDataPollTransmissions = 15;
 
     static constexpr uint8_t kReassemblyTimeout      = OPENTHREAD_CONFIG_6LOWPAN_REASSEMBLY_TIMEOUT; // in seconds.
-    static constexpr uint8_t kMeshHeaderFrameMtu     = OT_RADIO_FRAME_MAX_SIZE; // Max MTU with a Mesh Header frame.
+    static constexpr uint16_t kMeshHeaderFrameMtu     = OT_RADIO_FRAME_MAX_SIZE; // Max MTU with a Mesh Header frame.
     static constexpr uint8_t kMeshHeaderFrameFcsSize = sizeof(uint16_t);        // Frame FCS size for Mesh Header frame.
 
     // Hops left to use in lowpan mesh header: We use `kMaxRouteCost` as
